Create CSS document
The first steps in the document were to define the background of the
body to black. Likewise, the main container was centered by means of
flexbox. On the other hand, the default padding and margin of the body
were changed to 0.

Lastly, the container was defined with a height of 100% and width of
65%.