Layout elements within main container
The elements were arranged by means of grid, where it was specified
to have 2 columns and 1 row.

The container for the imagen occupies the whole first column, whereas
the second column is made up of 3 subcontainers.

These other 3 subcontainers were arranged with another grid, in which
1 column was set and 3 rows.

All of the containers have a color white or close to it.