Link image in the HTML and CSS documents
The DIV element that was in the container for the image was replaced
by the element IMG to link the image.

In the CSS document the IMG was selected to specify that the
height and the width of the image to fit with the space defined
previously with CSS grid.