Style inputs and labels
The inputs have no borders, and the outline was specified as none.
On the other hand, each DIV containing the corresponding INPUT and
LABEL was positioned as relative and the labels as absolute, so the
labels could be inside the Inputs since they are going to act as
placeholders as well.