Style inputs with hover and add span in html file
The span element was added in between the Input and label elements
in order to emulate a bar charging whenever an Input was selected. The
pseudo element ::before was used to accomplish this.

On the other hand, the labels were fixed at a sepecific point, so the
only animation they have is turning blue when the corresponding input
is selected.

Finally, margin bottom was added to each input to optimize space.