Place logo at the top of the image
The container for the logo was placed by applying to it position
absolute and the container for the image with position relative.

The container for the logo has an opacity to prevent it from obscuring
the main image. Likewise, an external font was downloaded and applied it
as well as default fonts.